Stronger integration with Shiny
Most of the times Shiny communicates with leaflet via an object called proxy - it allows for interactive user changes on the map, storing "old" version of it in the memory, i.e. rendering the map from scratch is not necessary after each user's change.
The problem is that proxy object from leaflet package connects only to leafletOutput, not leafglOutput, therefore glify polygons/points/lines were not visible on proxy object - that is actual map in shiny app. 
In order to overcome that problem I modified leafletOutput object. Two functions from leaflet package will be now covered by this modification.